Reduce number of doMany
